More work on the Expressions Refactoring routine
More refactoring done. Included a couple namespace changes for types and renamed everything to follow the PassThruExpression naming convention. This should allow things to be a little more clean in the end.  

Need to build a project or package for the TextTableGenerator. EVERYTHING uses it and I shouldn't have to be defining it all over the place really